+++ src/xdoctest/__init__.py
@@ -330,52 +330,6 @@ from xdoctest.runner import (doctest_module, doctest_callable,)
 from xdoctest.exceptions import (DoctestParseError, ExitTestException,
                                  MalformedDocstr,)
 
-
-def _parse_changelog(fpath):
-    """
-    Helper to parse the changelog for the version to verify versions agree.
-
-    CommandLine:
-        xdoctest -m xdoctest.__init__ _parse_changelog --dev
-
-    Example:
-        >>> # xdoctest: +REQUIRES(--dev)
-        >>> fpath = 'CHANGELOG.md'
-        >>> _parse_changelog(fpath)
-    """
-    try:
-        from packaging.version import parse as LooseVersion
-    except ImportError:
-        from distutils.version import LooseVersion
-    import re
-    pat = re.compile(r'#.*Version ([0-9]+\.[0-9]+\.[0-9]+)')
-    # We can statically modify this to a constant value when we deploy
-    versions = []
-    with open(fpath, 'r') as file:
-        for line in file.readlines():
-            line = line.rstrip()
-            if line:
-                parsed = pat.search(line)
-                if parsed:
-                    print('parsed = {!r}'.format(parsed))
-                    try:
-                        version_text = parsed.groups()[0]
-                        version = LooseVersion(version_text)
-                        versions.append(version)
-                    except Exception:
-                        print('Failed to parse = {!r}'.format(line))
-
-    import pprint
-    print('versions = {}'.format(pprint.pformat(versions)))
-    assert sorted(versions)[::-1] == versions
-
-    import xdoctest
-    changelog_version = versions[0]
-    module_version = LooseVersion(xdoctest.__version__)
-    print('changelog_version = {!r}'.format(changelog_version))
-    print('module_version = {!r}'.format(module_version))
-    assert changelog_version == module_version
-
 __all__ = ['DoctestParseError', 'ExitTestException', 'MalformedDocstr',
            'doctest_module', 'doctest_callable', 'utils', 'docstr',
            '__version__']